+++ skyfield/tests/test_against_horizons.py
@@ -16,7 +16,7 @@ ra_arcsecond = 24.0 / 360.0 / 60.0 / 60.0
 meter = 1.0 / AU_M
 
 def ts():
-    yield api.load.timescale(builtin=True)
+    yield api.load.timescale()
 
 def compare(value, expected_value, epsilon):
     if hasattr(value, 'shape') or hasattr(expected_value, 'shape'):